feat(tests): move tests to rootdir
