+++ packages/client-direct/tsconfig.json
@@ -3,7 +3,13 @@
     "compilerOptions": {
         "outDir": "dist",
         "rootDir": "src",
-        "types": ["vitest/globals", "supertest"]
+        "types": [
+            "vitest/globals",
+            "supertest"
+        ]
     },
-    "include": ["src/**/*.ts", "__test__/**/*.ts"]
+    "include": [
+        "src/**/*.ts",
+        "src/__tests__/**/*.ts"
+    ]
 }
